refactor(editor): extract slash menu state and commands into useSlashMenu hook
