+++ packages/web/src/components/editor/MilkdownEditor.tsx
@@ -1,26 +1,8 @@
-import {
-  IconBlockquote,
-  IconBold,
-  IconCode,
-  IconH1,
-  IconH2,
-  IconH3,
-  IconH4,
-  IconH5,
-  IconH6,
-  IconItalic,
-  IconLink,
-  IconList,
-  IconListCheck,
-  IconListNumbers,
-  IconPhoto,
-  IconStrikethrough,
-  IconTable,
-} from '@tabler/icons-react';
-import React, { useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { useCallback, useEffect, useMemo, useRef, useState } from 'react';
 import { useAwareness } from '../../hooks/useAwareness';
 import { useFloatingToolbar } from '../../hooks/useFloatingToolbar';
 import { useMilkdown } from '../../hooks/useMilkdown';
+import { useSlashMenu } from '../../hooks/useSlashMenu';
 import { useWikiLinkSuggestions } from '../../hooks/useWikiLinkSuggestions';
 import { authClient } from '../../lib/auth-client';
 import { getLogger } from '../../logger-init';
@@ -176,29 +158,6 @@ export function MilkdownEditor({
     closeSuggestions,
   } = useWikiLinkSuggestions(workspaceId, editorRef);
 
-  const [slashMenuState, setSlashMenuState] = useState({
-    isOpen: false,
-    query: '',
-    position: null as { x: number; y: number; top?: number; bottom?: number } | null,
-    range: null as { from: number; to: number } | null,
-  });
-
-  const handleSlashMenuSuggest = useCallback(
-    (
-      isOpen: boolean,
-      query: string,
-      position: { x: number; y: number; top?: number; bottom?: number } | null,
-      range: { from: number; to: number } | null,
-    ) => {
-      setSlashMenuState({ isOpen, query, position, range });
-    },
-    [],
-  );
-
-  const closeSlashMenu = useCallback(() => {
-    setSlashMenuState((prev) => ({ ...prev, isOpen: false }));
-  }, []);
-
   const [activeStates, setActiveStates] = useState({
     isBoldActive: false,
     isItalicActive: false,
@@ -339,6 +298,15 @@ export function MilkdownEditor({
     });
   }, [pageId, doc]);
 
+  const handleSlashMenuSuggestRef = useRef<
+    (
+      isOpen: boolean,
+      query: string,
+      position: { x: number; y: number; top?: number; bottom?: number } | null,
+      range: { from: number; to: number } | null,
+    ) => void
+  >(() => {});
+
   const { setContainer, editor } = useMilkdown({
     ...(initialValue !== undefined && { initialValue }),
     ...(onChange !== undefined && { onChange }),
@@ -346,7 +314,9 @@ export function MilkdownEditor({
     provider,
     onWikiLinkClick,
     onWikiLinkSuggest: handleWikiLinkSuggest,
-    onSlashMenuSuggest: handleSlashMenuSuggest,
+    onSlashMenuSuggest: useCallback((isOpen, query, position, range) => {
+      handleSlashMenuSuggestRef.current(isOpen, query, position, range);
+    }, []),
   });
 
   useAwareness(provider);
@@ -515,346 +485,6 @@ export function MilkdownEditor({
     });
   };
 
-  // Slash-specific handlers that don't call keepVisible()
-  const handleSlashParagraph = () => {
-    if (!editor) return;
-    editor.action((ctx) => {
-      const view = ctx.get(editorViewCtx);
-      if (!view) return;
-      const { state, dispatch } = view;
-      const paraType = state.schema.nodes.paragraph;
-      if (!paraType) return;
-      const command = setBlockType(paraType as never);
-      command(state, dispatch);
-    });
-  };
-
-  const handleSlashHeading = (level: number) => {
-    if (!editor) return;
-    editor.action((ctx) => {
-      const view = ctx.get(editorViewCtx);
-      if (!view) return;
-      const { state, dispatch } = view;
-      const headingType = state.schema.nodes.heading;
-      if (!headingType) return;
-      const command = setBlockType(headingType as never, { level });
-      command(state, dispatch);
-    });
-  };
-
-  const handleSlashQuote = () => {
-    if (!editor) return;
-    editor.action((ctx) => {
-      const view = ctx.get(editorViewCtx);
-      if (!view) return;
-      const { state, dispatch } = view;
-      const blockquoteType = state.schema.nodes.blockquote;
-      if (!blockquoteType) return;
-      const command = wrapIn(blockquoteType as never);
-      command(state, dispatch);
-    });
-  };
-
-  const handleSlashBulletList = () => {
-    if (!editor) return;
-    editor.action((ctx) => {
-      const view = ctx.get(editorViewCtx);
-      if (!view) return;
-      const { state, dispatch } = view;
-      const bulletListType = state.schema.nodes.bullet_list;
-      if (!bulletListType) return;
-      const command = wrapIn(bulletListType as never);
-      command(state, dispatch);
-    });
-  };
-
-  const handleSlashOrderedList = () => {
-    if (!editor) return;
-    editor.action((ctx) => {
-      const view = ctx.get(editorViewCtx);
-      if (!view) return;
-      const { state, dispatch } = view;
-      const orderedListType = state.schema.nodes.ordered_list;
-      if (!orderedListType) return;
-      const command = wrapIn(orderedListType as never);
-      command(state, dispatch);
-    });
-  };
-
-  const handleSlashTaskList = () => {
-    if (!editor) return;
-    editor.action((ctx) => {
-      const view = ctx.get(editorViewCtx);
-      if (!view) return;
-      const { state, dispatch } = view;
-      const bulletListType = state.schema.nodes.bullet_list;
-      const listItemType = state.schema.nodes.list_item;
-      if (!bulletListType || !listItemType) return;
-
-      const command = wrapIn(bulletListType as never);
-      command(state, dispatch);
-
-      const newState = view.state;
-      const tr = newState.tr;
-      const { from, to } = newState.selection;
-      newState.doc.nodesBetween(from, to, (node, pos) => {
-        if (node.type === listItemType && node.attrs.checked == null) {
-          tr.setNodeMarkup(pos, undefined, { ...node.attrs, checked: false });
-        }
-      });
-      if (tr.docChanged) {
-        dispatch(tr);
-      }
-    });
-  };
-
-  const handleSlashInsertTable = () => {
-    if (!editor) return;
-    editor.action((ctx) => {
-      const commands = ctx.get(commandsCtx);
-      commands.call(insertTableCommand.key, { row: 3, col: 3 });
-    });
-  };
-
-  const handleSlashDivider = () => {
-    if (!editor) return;
-    editor.action((ctx) => {
-      const view = ctx.get(editorViewCtx);
-      if (!view) return;
-      const { state, dispatch } = view;
-      const hrType = state.schema.nodes.hr;
-      if (!hrType) return;
-      const { $from } = state.selection;
-      const node = hrType.create();
-      const tr = state.tr.insert($from.pos, node);
-      dispatch(tr);
-    });
-  };
-
-  const handleSlashTag = () => {
-    if (!editor) return;
-    editor.action((ctx) => {
-      const view = ctx.get(editorViewCtx);
-      if (!view) return;
-      const { state, dispatch } = view;
-      const { $from } = state.selection;
-      const tr = state.tr.insertText('#tag ', $from.pos);
-      dispatch(tr);
-    });
-  };
-
-  const removeSlashTrigger = () => {
-    const range = slashMenuState.range;
-    if (!editor || !range) return;
-    editor.action((ctx) => {
-      const view = ctx.get(editorViewCtx);
-      if (!view) return;
-      const { state, dispatch } = view;
-      const { from, to } = range;
-      const tr = state.tr.delete(from, to);
-      const cursorPos = from;
-      tr.setSelection(TextSelection.near(tr.doc.resolve(cursorPos)));
-      dispatch(tr);
-    });
-  };
-
-  const executeSlashAction = (action: () => void) => {
-    removeSlashTrigger();
-    closeSlashMenu();
-    setTimeout(() => {
-      action();
-      if (editor) {
-        editor.action((ctx) => {
-          const view = ctx.get(editorViewCtx);
-          if (view) {
-            view.focus();
-          }
-        });
-      }
-    }, 0);
-  };
-
-  const slashCommands = [
-    {
-      id: 'paragraph',
-      label: 'Paragraph',
-      hint: 'P',
-      shortcut: 'Ctrl+Alt+0',
-      keywords: ['paragraph', 'text', 'p'],
-      icon: <span className="text-xs">¶</span>,
-      onSelect: () => executeSlashAction(handleSlashParagraph),
-    },
-    {
-      id: 'h1',
-      label: 'Heading 1',
-      hint: 'H1',
-      shortcut: 'Ctrl+Alt+1',
-      keywords: ['heading', 'h1', 'title'],
-      icon: <IconH1 size={16} />,
-      onSelect: () => executeSlashAction(() => handleSlashHeading(1)),
-    },
-    {
-      id: 'h2',
-      label: 'Heading 2',
-      hint: 'H2',
-      shortcut: 'Ctrl+Alt+2',
-      keywords: ['heading', 'h2'],
-      icon: <IconH2 size={16} />,
-      onSelect: () => executeSlashAction(() => handleSlashHeading(2)),
-    },
-    {
-      id: 'h3',
-      label: 'Heading 3',
-      hint: 'H3',
-      shortcut: 'Ctrl+Alt+3',
-      keywords: ['heading', 'h3'],
-      icon: <IconH3 size={16} />,
-      onSelect: () => executeSlashAction(() => handleSlashHeading(3)),
-    },
-    {
-      id: 'h4',
-      label: 'Heading 4',
-      hint: 'H4',
-      shortcut: 'Ctrl+Alt+4',
-      keywords: ['heading', 'h4'],
-      icon: <IconH4 size={16} />,
-      onSelect: () => executeSlashAction(() => handleSlashHeading(4)),
-    },
-    {
-      id: 'h5',
-      label: 'Heading 5',
-      hint: 'H5',
-      shortcut: 'Ctrl+Alt+5',
-      keywords: ['heading', 'h5'],
-      icon: <IconH5 size={16} />,
-      onSelect: () => executeSlashAction(() => handleSlashHeading(5)),
-    },
-    {
-      id: 'h6',
-      label: 'Heading 6',
-      hint: 'H6',
-      shortcut: 'Ctrl+Alt+6',
-      keywords: ['heading', 'h6'],
-      icon: <IconH6 size={16} />,
-      onSelect: () => executeSlashAction(() => handleSlashHeading(6)),
-    },
-    {
-      id: 'bold',
-      label: 'Bold',
-      hint: 'Bold',
-      shortcut: 'Ctrl+B',
-      keywords: ['bold', 'strong'],
-      icon: <IconBold size={16} />,
-      onSelect: () => executeSlashAction(handleBold),
-    },
-    {
-      id: 'italic',
-      label: 'Italic',
-      hint: 'Italic',
-      shortcut: 'Ctrl+I',
-      keywords: ['italic', 'emphasis'],
-      icon: <IconItalic size={16} />,
-      onSelect: () => executeSlashAction(handleItalic),
-    },
-    {
-      id: 'strikethrough',
-      label: 'Strikethrough',
-      hint: 'Strike',
-      shortcut: 'Ctrl+Shift+X',
-      keywords: ['strikethrough', 'strike'],
-      icon: <IconStrikethrough size={16} />,
-      onSelect: () => executeSlashAction(handleStrike),
-    },
-    {
-      id: 'code',
-      label: 'Code',
-      hint: 'Code',
-      shortcut: 'Ctrl+`',
-      keywords: ['code', 'inline'],
-      icon: <IconCode size={16} />,
-      onSelect: () => executeSlashAction(handleCode),
-    },
-    {
-      id: 'blockquote',
-      label: 'Blockquote',
-      hint: 'Quote',
-      shortcut: 'Ctrl+Shift+>',
-      keywords: ['quote', 'blockquote', 'citation'],
-      icon: <IconBlockquote size={16} />,
-      onSelect: () => executeSlashAction(handleSlashQuote),
-    },
-    {
-      id: 'link',
-      label: 'Link',
-      hint: 'Link',
-      shortcut: 'Ctrl+K',
-      keywords: ['link', 'url'],
-      icon: <IconLink size={16} />,
-      onSelect: () => executeSlashAction(handleLink),
-    },
-    {
-      id: 'bullet-list',
-      label: 'Bullet List',
-      hint: 'Bullet',
-      shortcut: 'Ctrl+Shift+8',
-      keywords: ['bullet', 'list', 'unordered'],
-      icon: <IconList size={16} />,
-      onSelect: () => executeSlashAction(handleSlashBulletList),
-    },
-    {
-      id: 'ordered-list',
-      label: 'Ordered List',
-      hint: 'Ordered',
-      shortcut: 'Ctrl+Shift+7',
-      keywords: ['ordered', 'list', 'number', 'numbered'],
-      icon: <IconListNumbers size={16} />,
-      onSelect: () => executeSlashAction(handleSlashOrderedList),
-    },
-    {
-      id: 'task-list',
-      label: 'Task List',
-      hint: 'Check',
-      shortcut: 'Ctrl+Shift+[',
-      keywords: ['task', 'check', 'list', 'todo', 'checkbox'],
-      icon: <IconListCheck size={16} />,
-      onSelect: () => executeSlashAction(handleSlashTaskList),
-    },
-    {
-      id: 'table',
-      label: 'Table',
-      hint: 'Table',
-      keywords: ['table', 'grid'],
-      icon: <IconTable size={16} />,
-      onSelect: () => executeSlashAction(handleSlashInsertTable),
-    },
-    {
-      id: 'image',
-      label: 'Image',
-      hint: 'Img',
-      shortcut: 'Ctrl+Shift+I',
-      keywords: ['image', 'photo', 'upload'],
-      icon: <IconPhoto size={16} />,
-      onSelect: () => executeSlashAction(handleImageUploadFromSlash),
-    },
-    {
-      id: 'divider',
-      label: 'Divider',
-      hint: 'Line',
-      keywords: ['divider', 'hr', 'line', 'separator', 'horizontal rule'],
-      icon: <span className="text-lg">—</span>,
-      onSelect: () => executeSlashAction(handleSlashDivider),
-    },
-    {
-      id: 'tag',
-      label: 'Tag',
-      hint: 'Tag',
-      shortcut: 'Ctrl+Shift+#',
-      keywords: ['tag', 'label', 'property', '#'],
-      icon: <span className="text-sm">#</span>,
-      onSelect: () => executeSlashAction(handleSlashTag),
-    },
-  ];
-
   const handleBold = () => {
     runMarkCommand('strong');
     setTimeout(updateActiveStates, 0);
@@ -969,6 +599,21 @@ export function MilkdownEditor({
     };
     input.click();
   };
+
+  const { slashMenuState, handleSlashMenuSuggest, closeSlashMenu, slashCommands } = useSlashMenu(
+    editorRef,
+    {
+      handleBold,
+      handleItalic,
+      handleStrike,
+      handleCode,
+      handleLink,
+      handleImageUploadFromSlash,
+    },
+  );
+
+  handleSlashMenuSuggestRef.current = handleSlashMenuSuggest;
+
   const handleH1 = () => {
     runBlockCommand('heading', { level: 1 });
     setTimeout(updateActiveStates, 0);
